chore: create agave-test-utilities crate and kill rpc dep to unified scheduler
