+++ client-test/Cargo.toml
@@ -43,6 +43,7 @@ tungstenite = { workspace = true, features = ["rustls-tls-webpki-roots"] }
 
 [dev-dependencies]
 agave-logger = { path = "../logger", features = ["agave-unstable-api"] }
+agave-test-utilities = { workspace = true, features = ["dev-context-only-utils"] }
 async-trait = { workspace = true }
 log = { workspace = true }
 solana-net-utils = { path = "../net-utils", features = ["agave-unstable-api"] }
