+++ client-test/tests/client.rs
@@ -1,4 +1,5 @@
 use {
+    agave_test_utilities::populate_blockstore_for_tests,
     futures_util::StreamExt,
     serde_json::{Value, json},
     solana_clock::Slot,
@@ -11,7 +12,7 @@ use {
     solana_pubsub_client::{nonblocking, pubsub_client::PubsubClient},
     solana_rpc::{
         optimistically_confirmed_bank_tracker::OptimisticallyConfirmedBank,
-        rpc::{create_test_transaction_entries, populate_blockstore_for_tests},
+        rpc::create_test_transaction_entries,
         rpc_pubsub_service::{PubSubConfig, PubSubService},
         rpc_subscriptions::RpcSubscriptions,
     },
