+++ core/Cargo.toml
@@ -192,6 +192,7 @@ sysctl = { workspace = true }
 [dev-dependencies]
 agave-reserved-account-keys = { path = "../reserved-account-keys", features = ["agave-unstable-api"] }
 agave-scheduler-bindings = { path = "../scheduler-bindings", features = ["agave-unstable-api"] }
+agave-test-utilities = { workspace = true, features = ["dev-context-only-utils"] }
 bencher = { workspace = true }
 bitvec = { workspace = true }
 criterion = { workspace = true }
