+++ rpc/Cargo.toml
@@ -22,10 +22,8 @@ name = "solana_rpc"
 agave-unstable-api = []
 dev-context-only-utils = [
     "agave-votor-messages/dev-context-only-utils",
-    "dep:solana-unified-scheduler-pool",
     "solana-ledger/dev-context-only-utils",
     "solana-rpc/dev-context-only-utils",
-    "solana-unified-scheduler-pool/dev-context-only-utils",
 ]
 
 [dependencies]
@@ -93,7 +91,6 @@ solana-transaction = { workspace = true }
 solana-transaction-context = { workspace = true }
 solana-transaction-error = { workspace = true }
 solana-transaction-status = { workspace = true }
-solana-unified-scheduler-pool = { workspace = true, optional = true }
 solana-validator-exit = { workspace = true }
 solana-version = { workspace = true }
 solana-vote = { workspace = true }
@@ -109,6 +106,7 @@ wincode = { workspace = true }
 
 [dev-dependencies]
 agave-reserved-account-keys = { path = "../reserved-account-keys", features = ["agave-unstable-api"] }
+agave-test-utilities = { workspace = true, features = ["dev-context-only-utils"] }
 serial_test = { workspace = true }
 solana-address-lookup-table-interface = { workspace = true }
 solana-client = { workspace = true }
